[Mac] Port to Xamarin.Mac 2.0

+++ Xwt.Mac/Xwt.Mac/AlertDialogBackend.cs
@@ -25,8 +25,8 @@
 // THE SOFTWARE.
 
 using System;
-using MonoMac.AppKit;
-using MonoMac.Foundation;
+using AppKit;
+using Foundation;
 using Xwt.Backends;
 using Xwt.Drawing;
 
@@ -57,7 +57,7 @@ namespace Xwt.Mac
 				this.AddButton (cmd.Label);
 			}
 
-			return message.Buttons [this.RunModal () - 1000];
+			return message.Buttons [(int)this.RunModal () - 1000];
 		}
 
 		public bool ApplyToAll { get; set; }
