+++ Xwt.Mac/Xwt.Mac/NotebookBackend.cs
@@ -25,7 +25,7 @@
 // THE SOFTWARE.
 
 using System;
-using MonoMac.AppKit;
+using AppKit;
 using Xwt.Backends;
 using System.Collections.Generic;
 
@@ -97,7 +97,7 @@ namespace Xwt.Mac
 		
 		public int CurrentTab {
 			get {
-				return Widget.IndexOf (Widget.Selected);
+				return (int) Widget.IndexOf (Widget.Selected);
 			}
 			set {
 				Widget.SelectAt (value);
